resolved all issues
